Former-commit-id: 3c800d67e52d3ce4ee7df7ff7ac0ec82018e1c6a

+++ src/gameplayer/front_end/gui_generator/button_generator/AbstractButton.java
@@ -6,7 +6,7 @@ import javafx.scene.input.MouseEvent;
 
 public class AbstractButton implements IButton {
 	
-	public Button createButton(String aMessage, int aXPos, int aYPos, 
+	public Button createButton(String aMessage, double aXPos, double aYPos, 
 		EventHandler<? super MouseEvent> aHandler){
 		Button newButton = new Button();
 		newButton.setOnMouseClicked(aHandler);
